+++ hyperfit-okhttp2-client/src/main/java/org/hyperfit/net/okhttp2/OkHttp2HyperClient.java
@@ -142,12 +142,12 @@ public class OkHttp2HyperClient extends BaseHyperClient {
             .addRequest(request)
         ;
 
-        Headers headers = response.headers();
-        for (String headerName : headers.names()) {
-            for(String value : headers.values(headerName)) {
-                responseBuilder.addHeader(headerName, value);
+        for( Entry<String,java.util.List<String>> h  :response.headers().toMultimap().entrySet()){
+            for(String val : h.getValue()){
+                responseBuilder.addHeader(h.getKey(), val);
             }
         }
+
 
         //Set the content type explicitly, even though it comes from the headers.  Hyperfit needs to know this
         //abstracted from the headers